+++ src/pages/AdminBooksPage/AdminBooksPage.css
@@ -12,6 +12,18 @@
   padding-bottom: 200px;
 }
 
+.button-div {
+  padding-bottom: 1%;
+  padding-right: 1%;
+  display: flex;
+  justify-content: end;
+}
+
+.button-div-mobile {
+  padding-bottom: 1%;
+  padding-left: 1%;
+}
+
 .block {
   margin-left: 270px;
   box-shadow: 0px 1px 4px rgba(59, 59, 59, 0.1);
