+++ src/pages/AdminBooksPage/index.tsx
@@ -4,6 +4,8 @@ import SideMenu from "../../components/SideMenu";
 import BooksContextProvider from "../../store/books/catalog/BooksContext";
 import BooksTable from "./BooksTable";
 import { MediaQueries } from "../../constants/mediaQueries";
+import NewBookButton from "./NewBookButton";
+import NewBookForm from "../../components/NewBookForm";
 import AdminNavBar from "../../components/AdminNavBar";
 
 const AdminBooksPage = () => {
@@ -14,7 +16,11 @@ const AdminBooksPage = () => {
       <SideMenu />
       <AdminNavBar />
       <div className={isLargeScreen ? "main" : "main-mobile"}>
+        <div className={isLargeScreen ? "button-div" : "button-div-mobile"}>
+          <NewBookButton />
+        </div>
         <BooksContextProvider>
+          <NewBookForm />
           <BooksTable />
         </BooksContextProvider>
       </div>
